1.3.6 - fix alias download loop & minor tweaks

+++ AutoUpdatingPlugin/LocalFileInspection/DllFileChecker.cs
@@ -50,15 +50,29 @@ namespace AutoUpdatingPlugin
 
 							modName = melonInfoAttribute.ConstructorArguments[1].Value as string;
 							modVersion = melonInfoAttribute.ConstructorArguments[2].Value as string;
+
+							//Logger.Msg($"found '{modName}' > '{modVersion}' > {assembly.Name.Version.ToString()}");
+
+							if ((VersionData)assembly.Name.Version.ToString() != (VersionData)modVersion)
+							{
+								Logger.Warning($"Version MisMatch: {modName} MelonInfo = {modVersion} Assembly = {assembly.Name.Version.ToString()}");
+							}
+							if ((VersionData)assembly.Name.Version.ToString() > (VersionData)modVersion)
+							{
+								modVersion = assembly.Name.Version.ToString();
+								Logger.Warning($"Using Assembly Version: {modName} {assembly.Name.Version.ToString()}");
+							}
 						}
 
 						bool isAliasName = APIList.IsAliasName(modName);
+						string newName = APIList.GetNewModName(modName);
 
 						if (isAliasName)
 						{
-							string newName = APIList.GetNewModName(modName);
-							Logger.Msg($"'{modName}' is obsolete. It will be replaced with '{newName}'.");
-							modName = newName;
+							if (!InstalledModList.installedMods.ContainsKey(newName)) {
+								Logger.Msg($"'{modName}' is obsolete. It will be replaced with '{newName}'.");
+								modName = newName;
+							}
 						}
 
 						modName = FileUtils.GetCleanName(modName);
@@ -68,17 +82,23 @@ namespace AutoUpdatingPlugin
 							if (installedModDetail.files[0].version > (VersionData)modVersion)
 							{
 								File.Delete(filename); // Delete duplicated mods
-								Logger.Msg("Deleted duplicated mod " + modName);
+								Logger.Msg("1Deleted duplicated mod " + modName);
 							}
 							else
 							{
-								File.Delete(installedModDetail.files[0].filepath); // Delete duplicated mods
-								installedModDetail.files.RemoveAt(0);
-								Logger.Msg("Deleted duplicated mod " + modName);
-								installedModDetail.files.Add(new InstalledFileDetail(modName, modVersion, filename, InstalledFileType.DLL));
-								if (isAliasName)
+								if (!InstalledModList.installedMods.ContainsKey(newName))
 								{
-									installedModDetail.TriggerOutdated();
+									File.Delete(installedModDetail.files[0].filepath); // Delete duplicated mods
+									installedModDetail.files.RemoveAt(0);
+									Logger.Msg("2Deleted duplicated mod " + modName);
+									installedModDetail.files.Add(new InstalledFileDetail(modName, modVersion, filename, InstalledFileType.DLL));
+									if (isAliasName)
+									{
+										installedModDetail.TriggerOutdated();
+									}
+								} else
+								{
+									Logger.Warning($"Naming MisMatch: {modName} > {newName}");
 								}
 							}
 						}
